Add meta tag to fix media queries
The lack of the meta viewport tag causes issues with the multiple media queries that should be running in this file. Media queries after the first do not work when implemented. To test the before and after, edit the body-color property of the various media queries before adding the tag and use the Chrome device toolbar to test what happens when resizing down. Before the meta tag is added the smaller media queries won't change the color. Afterwards it will.